refactor(constant): 하드코딩 된 상수와 반복 사용되는 상수를 lottoStandard로 분리
lottoStandard에 LOTTO_BOUNDARY 추가
lottoStandard에 LOTTO_RANK 추가
