+++ src/constants/lottoStandard.js
@@ -1,3 +1,14 @@
 const MONEY_UNIT = 1000;
 
-export default MONEY_UNIT;
+const LOTTO_BOUNDARY = Object.freeze({ MIN: 1, MAX: 45 });
+
+const LOTTO_RANK = Object.freeze({
+  FIRST_RANK: 1,
+  SECOND_RANK: 2,
+  THIRD_RANK: 3,
+  FOURTH_RANK: 4,
+  FIFTH_RANK: 5,
+  SIXTH_RANK: 6,
+});
+
+export { MONEY_UNIT, LOTTO_BOUNDARY, LOTTO_RANK };
